tls: separate out cert validation from ContextImpl
