+++ source/extensions/quic_listeners/quiche/BUILD
@@ -107,6 +107,7 @@ envoy_cc_library(
         ":envoy_quic_proof_verifier_base_lib",
         ":envoy_quic_utils_lib",
         "//source/extensions/transport_sockets/tls:context_lib",
+        "//source/extensions/transport_sockets/tls/cert_validator:cert_validator_lib",
     ],
 )
